requirements.txt
vk-api==11.6.1
google-api-python-client==
google-auth-httplib2==
google-auth-oauthlib==
# --upgrade google-api-python-client google-auth-httplib2 google-auth-oauthlib